added thread test.
